modularised questions page & component
