+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export { Navbar } from "./Navbar/Navbar";
 export { Footer } from "./Footer/Footer";
+export { QuestionCard } from "./QuestionCard";
