+++ src/pages/QuestionsPage.tsx
@@ -1,0 +1,9 @@
+import { QuestionCard } from "../components";
+
+export const QuestionsPage = () => {
+  return (
+    <main className="quiz-wrapper">
+      <QuestionCard />
+    </main>
+  );
+};
